+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+import 'dart:isolate';
+
 import 'package:logging/logging.dart';
 import 'package:logging_appenders/logging_appenders.dart';
 
@@ -14,6 +16,15 @@ void main() {
   _logger.severe('This is bad.');
   _logger.shout('This is just impolite');
 
+  Isolate.run(_isolateWork, debugName: 'Child Isolate');
+  Isolate.run(_isolateWork);
+
   // optionally dispose of the appender.
   appender.dispose();
+}
+
+void _isolateWork() {
+  // Child isolates need to setup logging again.
+  PrintAppender.setupLogging();
+  _logger.fine('Hello world from child!');
 }