+++ lib/src/internal/print_appender_default.dart
@@ -1,9 +1,10 @@
 import 'package:logging/logging.dart';
+import 'package:logging_appenders/src/base_appender.dart';
 import 'package:logging_appenders/src/logrecord_formatter.dart';
 import 'package:logging_appenders/src/print_appender.dart';
 
 LogRecordFormatter defaultLogRecordFormatter() =>
-    const DefaultLogRecordFormatter();
+    BaseLogAppender.defaultFormatter();
 
 PrintAppender defaultCreatePrintAppender({
   LogRecordFormatter? formatter,